requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
discord==1.0.1
discord.py==1.3.4
ibm-cloud-sdk-core==1.5.1
ibm-watson==4.5.0
idna==2.10
multidict==4.7.6
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
urllib3==1.25.10
websocket-client==0.48.0
websockets==8.1
yarl==1.4.2
youtube-dl==2020.6.16.1
